Catch RuntimeException when method is too large
When javaslicer is instrumenting, if a method is too large it has to revert the
instrumentation and leave the method as it was. That procedure is written, but
the problem was that before that, a call to toByteArray from ClassWritter was
throwing RuntimeException and there was no catch to handle it.